fix(export): render mermaid diagrams in pdf
