+++ src/components/editor/preview.tsx
@@ -1,63 +1,7 @@
 import { useEffect, useRef, useState } from "react";
 import { ensureMarkdownReady, renderMarkdown, useTheme } from "@/lib";
 import inspectUrl from "@/assets/mascot/inspect.png";
-
-let mermaidLib: typeof import("mermaid")["default"] | null = null;
-let mermaidLoading: Promise<typeof import("mermaid")["default"]> | null = null;
-let mermaidInitializedTheme: "default" | "dark" | null = null;
-
-async function getMermaid(themeName: "default" | "dark") {
-  if (mermaidLib) return mermaidLib;
-  if (!mermaidLoading) {
-    mermaidLoading = import("mermaid").then((mod) => {
-      mod.default.initialize({
-        startOnLoad: false,
-        theme: themeName,
-        securityLevel: "strict",
-        fontFamily: "JetBrains Mono, ui-monospace, monospace",
-      });
-      mermaidInitializedTheme = themeName;
-      mermaidLib = mod.default;
-      return mod.default;
-    });
-  }
-  return mermaidLoading;
-}
-
-async function renderMermaidBlocks(root: HTMLElement, theme: "default" | "dark") {
-  const blocks = Array.from(root.querySelectorAll<HTMLPreElement>("pre.mdv-mermaid:not(.is-rendered)"));
-  if (blocks.length === 0) return;
-  const mermaid = await getMermaid(theme);
-  // only re-initialize when theme actually changed; avoids id-registry churn + flicker
-  if (mermaidInitializedTheme !== theme) {
-    mermaid.initialize({
-      startOnLoad: false,
-      theme,
-      securityLevel: "strict",
-      fontFamily: "JetBrains Mono, ui-monospace, monospace",
-    });
-    mermaidInitializedTheme = theme;
-  }
-
-  for (const pre of blocks) {
-    const code = pre.querySelector("code")?.textContent ?? "";
-    const id = pre.id || `mdv-mermaid-${Date.now()}-${Math.random().toString(36).slice(2, 6)}`;
-    try {
-      const { svg } = await mermaid.render(`${id}-svg`, code);
-      pre.innerHTML = svg;
-      pre.classList.add("is-rendered");
-    } catch (err) {
-      console.error("marka.md: mermaid render failed", err);
-      // safe fallback — build with textContent, never innerHTML on user input
-      pre.replaceChildren();
-      const codeEl = document.createElement("code");
-      codeEl.className = "mdv-mermaid__error";
-      codeEl.textContent = code;
-      pre.appendChild(codeEl);
-      pre.classList.add("is-rendered", "is-error");
-    }
-  }
-}
+import { renderMermaidBlocks } from "@/lib/mermaid";
 
 type PreviewProps = {
   source: string;
